feat: notification hardening
