+++ src/core/agent/services/connectionService.ts
@@ -390,7 +390,7 @@ class ConnectionService extends AgentService {
         id: operation.name,
         recordType: OperationPendingRecordType.Oobi,
       });
-      Agent.agent.signifyNotifications.addPendingOperationToQueue(
+      Agent.agent.keriaNotifications.addPendingOperationToQueue(
         pendingOperation
       );
     } else if (!operation.done) {
