+++ src/core/agent/services/keriaNotificationService.test.ts
@@ -5,7 +5,7 @@ import { OperationPendingRecord } from "../records/operationPendingRecord";
 import { ConnectionHistoryType } from "./connection.types";
 import { CredentialStatus } from "./credentialService.types";
 import { EventService } from "./eventService";
-import { SignifyNotificationService } from "./signifyNotificationService";
+import { KeriaNotificationService } from "./keriaNotificationService";
 
 const identifiersListMock = jest.fn();
 const identifiersGetMock = jest.fn();
@@ -19,7 +19,6 @@ const groupGetRequestMock = jest.fn();
 const oobiResolveMock = jest.fn();
 const queryKeyStateMock = jest.fn();
 const markNotificationMock = jest.fn();
-const deleteNotificationMock = jest.fn((id: string) => Promise.resolve(id));
 const getCredentialMock = jest.fn();
 const admitMock = jest.fn();
 const submitAdmitMock = jest.fn();
@@ -89,49 +88,6 @@ const credentialMetadataMock = {
   updatedAt: "2024-08-09T04:21:19.695Z",
 };
 
-const grantIpexMessageMock = {
-  exn: {
-    v: "KERI10JSON000516_",
-    t: "exn",
-    d: "EJ1jbI8vTFCEloTfSsZkBpV0bUJnhGVyak5q-5IFIglL",
-    i: "EC9bQGHShmp2Juayqp0C5XcheBiHyc1p54pZ_Op-B95x",
-    p: "",
-    dt: "2024-07-30T04:19:55.801000+00:00",
-    r: "/ipex/grant",
-    q: {},
-    a: {
-      m: "",
-      i: "EE-gjeEni5eCdpFlBtG7s4wkv7LJ0JmWplCS4DNQwW2G",
-    },
-    e: {
-      acdc: {
-        d: "EAe_JgQ636ic-k34aUQMjDFPp6Zd350gEsQA6HePBU5W",
-        i: "EC9bQGHShmp2Juayqp0C5XcheBiHyc1p54pZ_Op-B95x",
-        s: "EBIFDhtSE0cM4nbTnaMqiV1vUIlcnbsqBMeVMmeGmXOu",
-        a: {
-          d: "ELHCh_X2aw7C-aYesOM4La23a5lsoNuJDuCsJuxwO2nq",
-          i: "EE-gjeEni5eCdpFlBtG7s4wkv7LJ0JmWplCS4DNQwW2G",
-          dt: "2024-07-30T04:19:55.348000+00:00",
-          attendeeName: "ccc",
-        },
-      },
-      iss: {
-        t: "iss",
-        d: "EHStOgwJku_Ln-YN2ohgWUH-CI07SyJnFppSbF8kG4PO",
-        i: "EEqfWy-6jx_FG0RNuNxZBh_jq6Lq1OPuvX5m3v1Bzxdn",
-        s: "0",
-        dt: "2024-07-30T04:19:55.348000+00:00",
-      },
-      d: "EKBPPnWxYw2I5CtQSyhyn5VUdSTJ61qF_-h-NwmFRkIF",
-    },
-  },
-  pathed: {
-    acdc: "-IABEEqfWy-6jx_FG0RNuNxZBh_jq6Lq1OPuvX5m3v1Bzxdn0AAAAAAAAAAAAAAAAAAAAAAAEHStOgwJku_Ln-YN2ohgWUH-CI07SyJnFppSbF8kG4PO",
-    iss: "-VAS-GAB0AAAAAAAAAAAAAAAAAAAAAAAEEO0xKzC8FOAXV-JgFZGgb0aIT2A3cPXPt9_0l_qcGM9",
-    anc: "-AABAACBlQqbI_qNpKYkzIog6tauSgt0XufBvGtrumfbnhSInFjSwnaIqZi353QT-c1W_gE9KIz3rgX5QNNWLcqA7bcM",
-  },
-};
-
 const exchangesGetMock = jest.fn();
 const signifyClient = jest.mocked({
   connect: jest.fn(),
@@ -166,7 +122,6 @@ const signifyClient = jest.mocked({
   notifications: () => ({
     list: jest.fn(),
     mark: markNotificationMock,
-    deleteById: deleteNotificationMock,
   }),
   ipex: () => ({
     admit: admitMock,
@@ -248,7 +203,8 @@ const operationPendingGetAllMock = jest.fn();
 const operationPendingStorage = jest.mocked({
   save: jest.fn(),
   delete: jest.fn(),
-  deleteById: jest.fn((id: string) => Promise.resolve(id)),
+  deleteById: jest.fn(),
+  update: jest.fn(),
   findById: jest.fn(),
   findAllByQuery: jest.fn(),
   getAll: operationPendingGetAllMock,
@@ -263,14 +219,21 @@ const credentialStorage = jest.mocked({
   getCredentialMetadatasById: jest.fn(),
 });
 
-const signifyNotificationService = new SignifyNotificationService(
+const keriaNotificationService = new KeriaNotificationService(
   agentServicesProps,
   notificationStorage as any,
   identifierStorage as any,
   operationPendingStorage as any,
   connectionStorage as any,
   ipexMessageStorage as any,
-  credentialStorage as any
+  credentialStorage as any,
+  Agent.agent.basicStorage,
+  Agent.agent.multiSigs,
+  Agent.agent.ipexCommunications,
+  Agent.agent.identifiers,
+  Agent.agent.getKeriaOnlineStatus,
+  Agent.agent.markAgentStatus,
+  Agent.agent.connect
 );
 
 jest.mock("../../../core/agent/agent", () => ({
@@ -285,17 +248,11 @@ jest.mock("../../../core/agent/agent", () => ({
       ipexCommunications: {
         grantAcdcFromAgree: jest.fn(),
         createLinkedIpexMessageRecord: jest.fn(),
-        acceptAcdcFromMultisigExn: jest.fn(),
         markAcdc: jest.fn(),
       },
-      signifyNotifications: {
-        deleteNotificationRecordById: (id: string) =>
-          deleteNotificationMock(id),
+      keriaNotifications: {
         addPendingOperationToQueue: jest.fn(),
         markAcdcComplete: jest.fn(),
-      },
-      identifiers: {
-        getIdentifier: jest.fn(),
       },
     },
   },
@@ -372,7 +329,7 @@ describe("Signify notification service of agent", () => {
     groupGetRequestMock.mockResolvedValue([{ exn: { a: { gid: "id" } } }]);
     jest.useFakeTimers();
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(notificationStorage.save).toBeCalledTimes(2);
     expect(callback).toBeCalledTimes(2);
@@ -413,7 +370,7 @@ describe("Signify notification service of agent", () => {
       done: true,
     });
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(admitMock).toBeCalledTimes(1);
     expect(submitAdmitMock).toBeCalledTimes(1);
@@ -426,20 +383,20 @@ describe("Signify notification service of agent", () => {
     };
 
     notificationStorage.findById = jest.fn().mockResolvedValue(notification);
-    await signifyNotificationService.readNotification(notification.id);
+    await keriaNotificationService.readNotification(notification.id);
     expect(notificationStorage.update).toBeCalledTimes(1);
   });
 
   test("Should throw error when read an invalid notification", async () => {
     notificationStorage.findById = jest.fn().mockResolvedValue(null);
     await expect(
-      signifyNotificationService.readNotification("not-exist-noti-id")
-    ).rejects.toThrowError(SignifyNotificationService.NOTIFICATION_NOT_FOUND);
+      keriaNotificationService.readNotification("not-exist-noti-id")
+    ).rejects.toThrowError(KeriaNotificationService.NOTIFICATION_NOT_FOUND);
   });
 
   test("can delete keri notification by ID", async () => {
     const id = "uuid";
-    await signifyNotificationService.deleteNotificationRecordById(
+    await keriaNotificationService.deleteNotificationRecordById(
       id,
       NotificationRoute.ExnIpexGrant
     );
@@ -449,7 +406,7 @@ describe("Signify notification service of agent", () => {
 
   test("Should not mark local notification when we delete notification", async () => {
     const id = "uuid";
-    await signifyNotificationService.deleteNotificationRecordById(
+    await keriaNotificationService.deleteNotificationRecordById(
       id,
       NotificationRoute.LocalAcdcRevoked
     );
@@ -483,7 +440,7 @@ describe("Signify notification service of agent", () => {
       },
     ];
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(callback).toBeCalledTimes(0);
   });
@@ -506,7 +463,7 @@ describe("Signify notification service of agent", () => {
       },
     ];
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(callback).toBeCalledTimes(0);
   });
@@ -539,7 +496,7 @@ describe("Signify notification service of agent", () => {
       },
     ];
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(callback).toBeCalledTimes(0);
   });
@@ -551,15 +508,15 @@ describe("Signify notification service of agent", () => {
     };
 
     notificationStorage.findById = jest.fn().mockResolvedValue(notification);
-    await signifyNotificationService.unreadNotification(notification.id);
+    await keriaNotificationService.unreadNotification(notification.id);
     expect(notificationStorage.update).toBeCalledTimes(1);
   });
 
   test("Should throw error when unread an invalid notification", async () => {
     notificationStorage.findById = jest.fn().mockResolvedValue(null);
     await expect(
-      signifyNotificationService.unreadNotification("not-exist-noti-id")
-    ).rejects.toThrowError(SignifyNotificationService.NOTIFICATION_NOT_FOUND);
+      keriaNotificationService.unreadNotification("not-exist-noti-id")
+    ).rejects.toThrowError(KeriaNotificationService.NOTIFICATION_NOT_FOUND);
   });
 
   test("Should skip if there is a missing multi-sig identifier", async () => {
@@ -580,7 +537,7 @@ describe("Signify notification service of agent", () => {
       },
     ];
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(callback).toBeCalledTimes(0);
   });
@@ -630,7 +587,7 @@ describe("Signify notification service of agent", () => {
     };
     groupGetRequestMock.mockResolvedValue([multisigNotificationExn]);
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(markNotificationMock).toBeCalledWith(notes[0].i);
     expect(Agent.agent.multiSigs.joinAuthorization).toBeCalledTimes(1);
@@ -651,7 +608,7 @@ describe("Signify notification service of agent", () => {
       },
     ];
     getIpexMessageMetadataMock.mockResolvedValueOnce({});
-    await signifyNotificationService.processNotification(
+    await keriaNotificationService.processNotification(
       notification[0],
       callback
     );
@@ -683,10 +640,7 @@ describe("Signify notification service of agent", () => {
       new Error(IpexMessageStorage.IPEX_MESSAGE_METADATA_RECORD_MISSING)
     );
 
-    await signifyNotificationService.processNotification(
-      notification,
-      callback
-    );
+    await keriaNotificationService.processNotification(notification, callback);
     expect(
       Agent.agent.ipexCommunications.createLinkedIpexMessageRecord
     ).toHaveBeenCalledWith(
@@ -723,13 +677,13 @@ describe("Signify notification service of agent", () => {
       name: "name",
       done: true,
     });
-    await signifyNotificationService.processNotification(
-      notification,
-      callback
+    const addPendingOperationToQueueSpy = jest.spyOn(
+      keriaNotificationService,
+      "addPendingOperationToQueue"
     );
-    expect(
-      Agent.agent.signifyNotifications.addPendingOperationToQueue
-    ).toBeCalledTimes(1);
+    await keriaNotificationService.processNotification(notification, callback);
+    expect(addPendingOperationToQueueSpy).toBeCalledTimes(1);
+    addPendingOperationToQueueSpy.mockRestore();
   });
 
   test("Should call createLinkedIpexMessageRecord with CREDENTIAL_REQUEST_PRESENT_AGREE", async () => {
@@ -752,10 +706,7 @@ describe("Signify notification service of agent", () => {
       new Error(IpexMessageStorage.IPEX_MESSAGE_METADATA_RECORD_MISSING)
     );
 
-    await signifyNotificationService.processNotification(
-      notification,
-      callback
-    );
+    await keriaNotificationService.processNotification(notification, callback);
     expect(
       Agent.agent.ipexCommunications.createLinkedIpexMessageRecord
     ).toHaveBeenCalledWith(
@@ -782,13 +733,14 @@ describe("Signify notification service of agent", () => {
     ];
 
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(markNotificationMock).toBeCalledTimes(1);
   });
 
   test("Should skip if notification route is /multisig/exn and the identifier is missing ", async () => {
     const callback = jest.fn();
+    Agent.agent.multiSigs.hasMultisig = jest.fn().mockResolvedValue(false);
     notificationStorage.findAllByQuery = jest.fn().mockResolvedValue([]);
     identifierStorage.getIdentifierMetadata = jest
       .fn()
@@ -810,7 +762,7 @@ describe("Signify notification service of agent", () => {
     identifierStorage.getIdentifierMetadata = jest.fn().mockResolvedValue({});
 
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(markNotificationMock).toBeCalledTimes(1);
   });
@@ -857,209 +809,9 @@ describe("Signify notification service of agent", () => {
     getCredentialMock.mockResolvedValue(acdc);
 
     for (const notif of notes) {
-      await signifyNotificationService.processNotification(notif, callback);
+      await keriaNotificationService.processNotification(notif, callback);
     }
     expect(markNotificationMock).toBeCalledTimes(1);
-  });
-
-  test("Original grant is linked to first received /multisig/exn admit message, and no notification record is created", async () => {
-    const callback = jest.fn();
-    Agent.agent.identifiers.getIdentifier = jest
-      .fn()
-      .mockResolvedValueOnce(identifierMetadataRecordProps);
-
-    const notif = {
-      i: "string",
-      dt: "string",
-      r: false,
-      a: {
-        r: "/multisig/exn",
-        d: "string",
-        m: "",
-      },
-    };
-
-    exchangesGetMock
-      .mockResolvedValueOnce({
-        exn: {
-          v: "KERI10JSON00032d_",
-          t: "exn",
-          d: "ELW97_QXT2MWtsmWLCSR8RBzH-dcyF2gTJvt72I0wEFO",
-          i: "ECS7jn05fIP_JK1Ub4E6hPviRKEdC55QhxZToxDIHo_E",
-          rp: "EJ84hiNC0ts71HARE1ZkcnYAFJP0s-RiLNyzupnk7edn",
-          p: "",
-          dt: "2024-08-28T06:39:55.501000+00:00",
-          r: "/multisig/exn",
-          q: {},
-          a: {
-            i: "EJ84hiNC0ts71HARE1ZkcnYAFJP0s-RiLNyzupnk7edn",
-            gid: "EC1cyV3zLnGs4B9AYgoGNjXESyQZrBWygz3jLlRD30bR",
-          },
-          e: {
-            exn: {
-              v: "KERI10JSON000178_",
-              t: "exn",
-              d: "EKa94ERqArLOvNf9AmItMJtsoGKZPVb3e_pEo_1D37qt",
-              i: "EC1cyV3zLnGs4B9AYgoGNjXESyQZrBWygz3jLlRD30bR",
-              rp: "EJ84hiNC0ts71HARE1ZkcnYAFJP0s-RiLNyzupnk7edn",
-              p: "EAe_JgQ636ic-k34aUQMjDFPp6Zd350gEsQA6HePBU5W",
-              dt: "2024-08-28T06:39:51.416000+00:00",
-              r: "/ipex/admit",
-              q: {},
-              a: {
-                i: "EJ84hiNC0ts71HARE1ZkcnYAFJP0s-RiLNyzupnk7edn",
-                m: "",
-              },
-              e: {},
-            },
-            d: "EE8_Xc0ZUh_sUJLtmBpVSEr-RFS2mRUIpFyL-pmvtPvx",
-          },
-        },
-        pathed: {},
-      })
-      .mockResolvedValueOnce(grantIpexMessageMock);
-
-    getCredentialMock.mockResolvedValue(undefined);
-    notificationStorage.findAllByQuery = jest.fn().mockResolvedValue([
-      {
-        type: "NotificationRecord",
-        id: "id",
-        createdAt: new Date(),
-        a: {
-          r: "/exn/ipex/grant",
-          d: "EIDUavcmyHBseNZAdAHR3SF8QMfX1kSJ3Ct0OqS0-HCW",
-        },
-        route: "/exn/ipex/grant",
-        read: true,
-        linkedGroupRequests: {},
-        connectionId: "EEFjBBDcUM2IWpNF7OclCme_bE76yKE3hzULLzTOFE8E",
-        updatedAt: new Date(),
-      },
-    ]);
-
-    await signifyNotificationService.processNotification(notif, callback);
-
-    expect(notificationStorage.update).toBeCalledWith({
-      type: "NotificationRecord",
-      id: "id",
-      createdAt: new Date(),
-      a: {
-        r: "/exn/ipex/grant",
-        d: "EIDUavcmyHBseNZAdAHR3SF8QMfX1kSJ3Ct0OqS0-HCW",
-      },
-      route: "/exn/ipex/grant",
-      read: true,
-      linkedGroupRequests: {
-        "ELW97_QXT2MWtsmWLCSR8RBzH-dcyF2gTJvt72I0wEFO": false,
-      },
-      connectionId: "EEFjBBDcUM2IWpNF7OclCme_bE76yKE3hzULLzTOFE8E",
-      updatedAt: new Date(),
-    });
-    expect(markNotificationMock).toBeCalledTimes(1);
-    expect(notificationStorage.save).toBeCalledTimes(0);
-  });
-
-  test("Auto-joins /multisig/exn admit message and links to grant if we have joined a previous admit message, and no notification record is created", async () => {
-    const callback = jest.fn();
-    Agent.agent.identifiers.getIdentifier = jest
-      .fn()
-      .mockResolvedValueOnce(identifierMetadataRecordProps);
-
-    const noti = {
-      i: "string",
-      dt: "string",
-      r: false,
-      a: {
-        r: "/multisig/exn",
-        d: "string",
-        m: "",
-      },
-    };
-
-    exchangesGetMock
-      .mockResolvedValueOnce({
-        exn: {
-          v: "KERI10JSON00032d_",
-          t: "exn",
-          d: "ELW97_QXT2MWtsmWLCSR8RBzH-dcyF2gTJvt72I0wEFO",
-          i: "ECS7jn05fIP_JK1Ub4E6hPviRKEdC55QhxZToxDIHo_E",
-          rp: "EJ84hiNC0ts71HARE1ZkcnYAFJP0s-RiLNyzupnk7edn",
-          p: "",
-          dt: "2024-08-28T06:39:55.501000+00:00",
-          r: "/multisig/exn",
-          q: {},
-          a: {
-            i: "EJ84hiNC0ts71HARE1ZkcnYAFJP0s-RiLNyzupnk7edn",
-            gid: "EC1cyV3zLnGs4B9AYgoGNjXESyQZrBWygz3jLlRD30bR",
-          },
-          e: {
-            exn: {
-              v: "KERI10JSON000178_",
-              t: "exn",
-              d: "EKa94ERqArLOvNf9AmItMJtsoGKZPVb3e_pEo_1D37qt",
-              i: "EC1cyV3zLnGs4B9AYgoGNjXESyQZrBWygz3jLlRD30bR",
-              rp: "EJ84hiNC0ts71HARE1ZkcnYAFJP0s-RiLNyzupnk7edn",
-              p: "EAe_JgQ636ic-k34aUQMjDFPp6Zd350gEsQA6HePBU5W",
-              dt: "2024-08-28T06:39:51.416000+00:00",
-              r: "/ipex/admit",
-              q: {},
-              a: {
-                i: "EJ84hiNC0ts71HARE1ZkcnYAFJP0s-RiLNyzupnk7edn",
-                m: "",
-              },
-              e: {},
-            },
-            d: "EE8_Xc0ZUh_sUJLtmBpVSEr-RFS2mRUIpFyL-pmvtPvx",
-          },
-        },
-        pathed: {},
-      })
-      .mockResolvedValueOnce(grantIpexMessageMock);
-
-    getCredentialMock.mockResolvedValue(undefined);
-    notificationStorage.findAllByQuery = jest.fn().mockResolvedValue([
-      {
-        type: "NotificationRecord",
-        id: "id",
-        createdAt: new Date(),
-        a: {
-          r: "/exn/ipex/grant",
-          d: "EIDUavcmyHBseNZAdAHR3SF8QMfX1kSJ3Ct0OqS0-HCW",
-        },
-        route: "/exn/ipex/grant",
-        read: true,
-        linkedGroupRequests: {
-          "EDm8iNyZ9I3P93jb0lFtL6DJD-4Mtd2zw1ADFOoEQAqw": true,
-        },
-        connectionId: "EEFjBBDcUM2IWpNF7OclCme_bE76yKE3hzULLzTOFE8E",
-        updatedAt: new Date(),
-      },
-    ]);
-
-    await signifyNotificationService.processNotification(noti, callback);
-
-    expect(
-      Agent.agent.ipexCommunications.acceptAcdcFromMultisigExn
-    ).toBeCalledWith("ELW97_QXT2MWtsmWLCSR8RBzH-dcyF2gTJvt72I0wEFO");
-    expect(notificationStorage.update).toBeCalledWith({
-      type: "NotificationRecord",
-      id: "id",
-      createdAt: new Date(),
-      a: {
-        r: "/exn/ipex/grant",
-        d: "EIDUavcmyHBseNZAdAHR3SF8QMfX1kSJ3Ct0OqS0-HCW",
-      },
-      route: "/exn/ipex/grant",
-      read: true,
-      linkedGroupRequests: {
-        "ELW97_QXT2MWtsmWLCSR8RBzH-dcyF2gTJvt72I0wEFO": true,
-        "EDm8iNyZ9I3P93jb0lFtL6DJD-4Mtd2zw1ADFOoEQAqw": true,
-      },
-      connectionId: "EEFjBBDcUM2IWpNF7OclCme_bE76yKE3hzULLzTOFE8E",
-      updatedAt: new Date(),
-    });
-    expect(markNotificationMock).toBeCalledTimes(1);
-    expect(notificationStorage.save).toBeCalledTimes(0);
   });
 });
 
@@ -1092,10 +844,7 @@ describe("Long running operation tracker", () => {
     identifierStorage.getIdentifierMetadata.mockResolvedValueOnce({
       signifyName: "signifyName",
     });
-    await signifyNotificationService.processOperation(
-      operationRecord,
-      callback
-    );
+    await keriaNotificationService.processOperation(operationRecord, callback);
     expect(Agent.agent.multiSigs.endRoleAuthorization).toBeCalledWith(
       "signifyName"
     );
@@ -1124,10 +873,7 @@ describe("Long running operation tracker", () => {
       recordType: "witness",
       updatedAt: new Date("2024-08-01T10:36:17.814Z"),
     } as OperationPendingRecord;
-    await signifyNotificationService.processOperation(
-      operationRecord,
-      callback
-    );
+    await keriaNotificationService.processOperation(operationRecord, callback);
     expect(identifierStorage.updateIdentifierMetadata).toBeCalledWith(
       "AOCUvGbpidkplC7gAoJOxLgXX1P2j4xlWMbzk3gM8JzA",
       {
@@ -1165,10 +911,7 @@ describe("Long running operation tracker", () => {
       recordType: "oobi",
       updatedAt: new Date("2024-08-01T10:36:17.814Z"),
     } as OperationPendingRecord;
-    await signifyNotificationService.processOperation(
-      operationRecord,
-      callback
-    );
+    await keriaNotificationService.processOperation(operationRecord, callback);
     expect(connectionStorage.update).toBeCalledWith({
       id: connectionMock.id,
       pending: false,
@@ -1192,7 +935,6 @@ describe("Long running operation tracker", () => {
       .mockResolvedValueOnce({
         exn: {
           r: ExchangeRoute.IpexGrant,
-          d: "d",
           e: {
             acdc: {
               d: credentialIdMock,
@@ -1207,98 +949,11 @@ describe("Long running operation tracker", () => {
       recordType: "exchange.receivecredential",
       updatedAt: new Date("2024-08-01T10:36:17.814Z"),
     } as OperationPendingRecord;
-
-    identifierStorage.getIdentifierMetadata = jest.fn().mockResolvedValue({
-      type: "IdentifierMetadataRecord",
-      id: "EC1cyV3zLnGs4B9AYgoGNjXESyQZrBWygz3jLlRD30bR",
-      displayName: "holder",
-      signifyName: "764c965c-d997-4842-b940-aebd514fce42",
-      signifyOpName: "group.EC1cyV3zLnGs4B9AYgoGNjXESyQZrBWygz3jLlRD30bR",
-      createdAt: new Date(),
-      updatedAt: new Date(),
-    });
-
-    await signifyNotificationService.processOperation(
-      operationRecord,
-      callback
-    );
+    await keriaNotificationService.processOperation(operationRecord, callback);
     expect(Agent.agent.ipexCommunications.markAcdc).toBeCalledWith(
       credentialIdMock,
       CredentialStatus.CONFIRMED
     );
-    expect(notificationStorage.save).not.toBeCalled();
-    expect(operationPendingStorage.deleteById).toBeCalledTimes(1);
-  });
-
-  test("Should delete original grant notification when multi-sig admit operation completes", async () => {
-    const callback = jest.fn();
-    const credentialIdMock = "credentialId";
-    signifyClient
-      .exchanges()
-      .get.mockResolvedValueOnce({
-        exn: {
-          r: ExchangeRoute.IpexAdmit,
-          p: "p",
-        },
-      })
-      .mockResolvedValueOnce({
-        exn: {
-          r: ExchangeRoute.IpexGrant,
-          d: "d",
-          e: {
-            acdc: {
-              d: credentialIdMock,
-            },
-          },
-        },
-      });
-    const operationRecord = {
-      type: "OperationPendingRecord",
-      id: "exchange.receivecredential.AOCUvGbpidkplC7gAoJOxLgXX1P2j4xlWMbzk3gM8JzA",
-      createdAt: new Date("2024-08-01T10:36:17.814Z"),
-      recordType: "exchange.receivecredential",
-      updatedAt: new Date("2024-08-01T10:36:17.814Z"),
-    } as OperationPendingRecord;
-
-    identifierStorage.getIdentifierMetadata = jest.fn().mockResolvedValue({
-      type: "IdentifierMetadataRecord",
-      id: "EC1cyV3zLnGs4B9AYgoGNjXESyQZrBWygz3jLlRD30bR",
-      displayName: "holder",
-      signifyName: "764c965c-d997-4842-b940-aebd514fce42",
-      signifyOpName: "group.EC1cyV3zLnGs4B9AYgoGNjXESyQZrBWygz3jLlRD30bR",
-      multisigManageAid: "EAL7pX9Hklc_iq7pkVYSjAilCfQX3sr5RbX76AxYs2UH",
-      createdAt: new Date(),
-      updatedAt: new Date(),
-    });
-
-    notificationStorage.findAllByQuery = jest.fn().mockResolvedValue([
-      {
-        type: "NotificationRecord",
-        id: "id",
-        createdAt: new Date(),
-        a: {
-          r: "/exn/ipex/grant",
-          d: "EIDUavcmyHBseNZAdAHR3SF8QMfX1kSJ3Ct0OqS0-HCW",
-        },
-        route: "/exn/ipex/grant",
-        read: true,
-        linkedGroupRequests: {
-          "EDm8iNyZ9I3P93jb0lFtL6DJD-4Mtd2zw1ADFOoEQAqw": false,
-        },
-        connectionId: "EEFjBBDcUM2IWpNF7OclCme_bE76yKE3hzULLzTOFE8E",
-        updatedAt: new Date(),
-      },
-    ]);
-
-    await signifyNotificationService.processOperation(
-      operationRecord,
-      callback
-    );
-    expect(Agent.agent.ipexCommunications.markAcdc).toBeCalledWith(
-      credentialIdMock,
-      CredentialStatus.CONFIRMED
-    );
-    expect(notificationStorage.deleteById).toBeCalledWith("id");
     expect(operationPendingStorage.deleteById).toBeCalledTimes(1);
   });
 
@@ -1343,10 +998,7 @@ describe("Long running operation tracker", () => {
       recordType: "exchange.revokecredential",
       updatedAt: new Date("2024-08-01T10:36:17.814Z"),
     } as OperationPendingRecord;
-    await signifyNotificationService.processOperation(
-      operationRecord,
-      callback
-    );
+    await keriaNotificationService.processOperation(operationRecord, callback);
     expect(Agent.agent.ipexCommunications.markAcdc).toBeCalledWith(
       credentialIdMock,
       CredentialStatus.REVOKED
@@ -1399,10 +1051,7 @@ describe("Long running operation tracker", () => {
       recordType: "exchange.revokecredential",
       updatedAt: new Date("2024-08-01T10:36:17.814Z"),
     } as OperationPendingRecord;
-    await signifyNotificationService.processOperation(
-      operationRecord,
-      callback
-    );
+    await keriaNotificationService.processOperation(operationRecord, callback);
     expect(Agent.agent.ipexCommunications.markAcdc).not.toBeCalled();
     expect(
       Agent.agent.ipexCommunications.createLinkedIpexMessageRecord
@@ -1451,10 +1100,7 @@ describe("Long running operation tracker", () => {
       recordType: "exchange.revokecredential",
       updatedAt: new Date("2024-08-01T10:36:17.814Z"),
     } as OperationPendingRecord;
-    await signifyNotificationService.processOperation(
-      operationRecord,
-      callback
-    );
+    await keriaNotificationService.processOperation(operationRecord, callback);
     expect(Agent.agent.ipexCommunications.markAcdc).not.toBeCalled();
     expect(
       Agent.agent.ipexCommunications.createLinkedIpexMessageRecord
@@ -1495,10 +1141,7 @@ describe("Long running operation tracker", () => {
       recordType: "exchange.receivecredential",
       updatedAt: new Date("2024-08-01T10:36:17.814Z"),
     } as OperationPendingRecord;
-    await signifyNotificationService.processOperation(
-      operationRecord,
-      callback
-    );
+    await keriaNotificationService.processOperation(operationRecord, callback);
     expect(operationsGetMock).toBeCalledTimes(1);
     expect(Agent.agent.ipexCommunications.markAcdc).toBeCalledTimes(0);
     expect(operationPendingStorage.deleteById).toBeCalledTimes(1);
@@ -1521,7 +1164,7 @@ describe("Long running operation tracker", () => {
       },
     ]);
     try {
-      await signifyNotificationService.pollLongOperationsWithCb(callback);
+      await keriaNotificationService.pollLongOperationsWithCb(callback);
     } catch (error) {
       expect((error as Error).message).toBe("Force Exit");
     }
