+++ src/ui/App.test.tsx
@@ -53,7 +53,7 @@ jest.mock("../core/agent/agent", () => ({
         onBasicMessageStateChanged: jest.fn(),
         pickupMessagesFromMediator: jest.fn(),
       },
-      signifyNotifications: {
+      keriaNotifications: {
         pollNotificationsWithCb: jest.fn(),
         pollLongOperationsWithCb: jest.fn(),
         getAllNotifications: jest.fn(),
