+++ src/ui/pages/NotificationDetails/components/CredentialRequest/ChooseCredential.test.tsx
@@ -50,7 +50,7 @@ const offerAcdcFromApplyMock = jest.fn(
 jest.mock("../../../../../core/agent/agent", () => ({
   Agent: {
     agent: {
-      signifyNotifications: {
+      keriaNotifications: {
         deleteNotificationRecordById: (id: string) =>
           deleteNotificationMock(id),
       },
