+++ src/ui/pages/NotificationDetails/components/CredentialRequest/CredentialRequestInformation.tsx
@@ -46,7 +46,7 @@ const CredentialRequestInformation = ({
   };
 
   const handleDecline = async () => {
-    await Agent.agent.signifyNotifications.deleteNotificationRecordById(
+    await Agent.agent.keriaNotifications.deleteNotificationRecordById(
       notificationDetails.id,
       notificationDetails.a.r as NotificationRoute
     );
