+++ src/ui/pages/NotificationDetails/components/MultiSigRequest/MultiSigRequest.test.tsx
@@ -58,7 +58,7 @@ jest.mock("@ionic/react", () => ({
 jest.mock("../../../../../core/agent/agent", () => ({
   Agent: {
     agent: {
-      signifyNotifications: {
+      keriaNotifications: {
         deleteNotificationRecordById: (id: string) =>
           deleteNotificationMock(id),
       },
