+++ src/ui/pages/NotificationDetails/components/ReceiveCredential/ReceiveCredential.tsx
@@ -69,7 +69,7 @@ const ReceiveCredential = ({
   };
 
   const handleDecline = async () => {
-    await Agent.agent.signifyNotifications.deleteNotificationRecordById(
+    await Agent.agent.keriaNotifications.deleteNotificationRecordById(
       notificationDetails.id,
       notificationDetails.a.r as NotificationRoute
     );
