+++ src/ui/pages/Notifications/Notifications.test.tsx
@@ -26,7 +26,7 @@ jest.mock("../../../core/agent/agent", () => ({
           },
         }),
       },
-      signifyNotifications: {
+      keriaNotifications: {
         readNotification: (id: string) => readNotificationMock(id),
       },
       basicStorage: {
