+++ src/ui/pages/Notifications/components/NotificationOptionsModal.test.tsx
@@ -37,7 +37,7 @@ const unreadNotificationMock = jest.fn((id: string) => Promise.resolve(id));
 jest.mock("../../../../core/agent/agent", () => ({
   Agent: {
     agent: {
-      signifyNotifications: {
+      keriaNotifications: {
         deleteNotificationRecordById: (id: string) =>
           deleteNotificationMock(id),
         readNotification: (id: string) => readNotificationMock(id),
